change to uint16_t for channel addressing use.

+++ lib/base/datastruct/SLookup.h
@@ -49,7 +49,7 @@ struct SIFI_EXPORT SLookupChannel
     virtual ~SLookupChannel() = default;
 
     ///@{
-    uint8_t m, l, s; ///< component of virtual address
+    uint16_t m, l, s; ///< component of virtual address
     ///}@
 
     virtual uint read(const char* buffer);
